backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
python-dotenv
